Make minor changes in directory structure
